[scaleway] Add SSH KeyPair Support
It's not documented in their API docs, but it is used in their Golang CLI client, so I've worked out how to add SSH keys to a Scaleway account. It does require switching to the `account.scaleway.com` API endpoint, but the `PATCH` request does successfully replace the list of public keys belonging to the user. The entire list is then installed on each new server.